Fix match

+++ src/overlays/gname/gname.c
@@ -1400,7 +1400,7 @@ void* func_80142220(void* arg0, s32* arg1)
 /**
  * decomp.me (88.88%) https://decomp.me/scratch/UHlWz
  */
-void* func_80142274(void* arg0, s32* arg1, s32 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7)
+void* func_80142274(void* arg0, s32* arg1, u8 arg2, s32 arg3, s32 arg4, s32 arg5, s32 arg6, s32 arg7)
 {
     unsigned char* base = (unsigned char*)arg0;
     unsigned char* ptr = base;
